fixed SimpleBankAccountWithAtm

+++ pps-lab01-intellij-basic-example/src/lab01/example/model/SimpleBankAccountWithAtm.java
@@ -2,7 +2,7 @@ package lab01.example.model;
 
 public class SimpleBankAccountWithAtm extends AbstractSimpleBankAccount {
 
-    private static final int FEE = 1;
+    private static final double FEE = 1;
 
     public SimpleBankAccountWithAtm(final AccountHolder holder, final double balance) {
         super(balance, holder);
